% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrIO.R
\docType{package}
\name{learnrIO-package}
\alias{learnrIO}
\alias{learnrIO-package}
\title{learnrIO: An Interactive R Programming and Statistics Course for Industrial Organizational Psychology Graduate and Undergraduate Students}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ashleigh Wilson \email{ash.wil.1990@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
