% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrIO.R
\docType{package}
\name{learnrIO-package}
\alias{learnrIO}
\alias{learnrIO-package}
\title{learnrIO: An Interactive R Programming and Statistics Course for Industrial Organizational Psychology Graduate and Undergraduate Students}
\description{
An interactive R course for IO Psych graduate and undergraduate students.
}
\author{
\strong{Maintainer}: Ashleigh Wilson \email{ash.wil.1990@gmail.com}

Authors:
\itemize{
  \item Matthew Bennett \email{matthew.bennett@shl.com}
}

}
\keyword{internal}
